Move archive command to domain layer
