+++ lib/pod/commands.rb
@@ -4,12 +4,13 @@ require_relative "commands/base_runner"
 require_relative "commands/base_output"
 require_relative "commands/add/runner"
 require_relative "commands/add/output"
+require_relative "commands/archive/runner"
+require_relative "commands/archive/output"
 
 require_relative "commands/init"
 require_relative "commands/podcasts"
 require_relative "commands/episodes"
 require_relative "commands/open"
-require_relative "commands/archive"
 require_relative "commands/dearchive"
 require_relative "commands/delete"
 require_relative "commands/sync"
