+++ lib/pod/outputs/text.rb
@@ -4,7 +4,6 @@ require_relative "text/init"
 require_relative "text/podcasts"
 require_relative "text/episodes"
 require_relative "text/open"
-require_relative "text/archive"
 require_relative "text/dearchive"
 require_relative "text/delete"
 require_relative "text/sync"
