+++ spec/pod/commands/episodes_spec.rb
@@ -82,7 +82,7 @@ RSpec.describe Pod::Commands::Episodes do
 
     context "when there are archived episodes", :populate_db do
       it "ignores them" do
-        Pod::Commands::Archive.call(3)
+        Pod::Commands::Archive::Runner.call(3)
         result = described_class.call(1)
 
         expect(result[:status]).to eq(:success)
@@ -102,7 +102,7 @@ RSpec.describe Pod::Commands::Episodes do
 
     context "when there are archived episodes, but the all option is used", :populate_db do
       it "returns all episodes, including the archived ones" do
-        Pod::Commands::Archive.call(3)
+        Pod::Commands::Archive::Runner.call(3)
         result = described_class.call(1, {"all" => true})
 
         expect(result[:status]).to eq(:success)
